requirements.txt
streamlit==1.42.2
requests==2.32.3
duckdb==1.1.3
pillow==11.1.0
